src/third_party/blink/web_tests/virtual/web-components-v0-disabled/fast/dom/shadow/distribution-for-detached-subtree-expected.txt

CONSOLE ERROR: line 8: Uncaught TypeError: shadowHost.createShadowRoot is not a function
distribution should be recalculated correctly for a detached subtree.

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".

FAIL successfullyParsed should be true. Was false.

TEST COMPLETE
